+++ lib/ugigi/parser.rb
@@ -8,13 +8,28 @@ module Ugigi
     end
 
     def fetch(args)
+      case args[:only]
+      when :sosowa
+        args = args.update(:sswp => 0, :compe => 0)
+      when :compe
+        args = args.update(:sswl => 0, :sswp => 0)
+      when :sswp
+        args = args.update(:sswl => 0, :compe => 0)
+      end
+      args.delete(:only) if args.has_key? :only
       params = Ugigi.serialize_parameter(args)
       page = @agent.get(URI.join(BASE_URL, params))
       data = JSON.parse(page.body)
       indexes = []
       data.each do |e|
         index = Index.new(e)
-        indexes << index
+        case args[:style]
+        when :sosowa
+          index.element = index.to_sosowa_index
+          indexes << index
+        else
+          indexes << index
+        end
       end
       return indexes
     end