+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3
+* Ugigi::Index#to_sosowa_indexを追加しました。sosowa gemと同じIndex出力を返します。
+* Ugigi#searchに:styleと:onlyパラメタが追加されました。例えば:style => :sosowaでsosowa gem形式で、:only => :sosowaで創想話のみを対象に検索出来ます。
+
 ## 0.2
 * element要素にアクセス出来るようにしました。